Tests/LibWeb/Text/expected/canvas/drawimage-canvas-then-resize-source.txt

B kept A's pre-resize content (red): true
A was reset to transparent black: true
